requirements.txt
bidict==0.23.1
blinker==1.8.2
click==8.1.7
dnspython==1.16.0
Flask==3.0.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
future==1.0.0
greenlet==3.0.3
gunicorn==20.1.0
h11==0.14.0
iniconfig==2.0.0
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python-headless==4.9.0.80
packaging==24.0
pluggy==1.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyserial==3.5
pytest==8.2.1
python-engineio==4.9.1
python-socketio==5.11.2
PyYAML==6.0.1
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.1
uWSGI==2.0.26
Werkzeug==3.0.3
wsproto==1.2.0
